+++ apps/web/src/app/(needLogin)/travel/plan/page.tsx
@@ -1,24 +1,24 @@
-'use client';
-import { useSearchParams } from 'next/navigation';
+// 'use client';
+// import { useSearchParams } from 'next/navigation';
 
-import { travelPlanStep1config } from '@/constants';
+// import { travelPlanStep1config } from '@/constants';
 
-import Step1Title from '../_components/Step1Title/Step1Title';
-import TravelPlanTemplete from './stepTemplete';
+// import Step1Title from '../_components/Step1Title/Step1Title';
+// import TravelPlanTemplete from './stepTemplete';
 
-const TravelPlan = () => {
-  const params = useSearchParams();
-  const search = params.get('stepId');
+// const TravelPlan = () => {
+//   const params = useSearchParams();
+//   const search = params.get('stepId');
 
-  let StepComponent;
+//   let StepComponent;
 
-  if (search == '0') {
-    StepComponent = <Step1Title config={travelPlanStep1config} />;
-  } else {
-    StepComponent = <TravelPlanTemplete search={search} />;
-  }
+//   if (search == '0') {
+//     StepComponent = <Step1Title config={travelPlanStep1config} />;
+//   } else {
+//     StepComponent = <TravelPlanTemplete search={search} />;
+//   }
 
-  return <div>{StepComponent}</div>;
-};
+//   return <div>{StepComponent}</div>;
+// };
 
-export default TravelPlan;
+// export default TravelPlan;
